+++ src/Jigoshop/Payment/OnDelivery.php
@@ -5,6 +5,7 @@ namespace Jigoshop\Payment;
 use Jigoshop\Core\Options;
 use Jigoshop\Entity\Order;
 use Jigoshop\Exception;
+use Jigoshop\Service\OrderServiceInterface;
 use WPAL\Wordpress;
 
 class OnDelivery implements Method
@@ -15,12 +16,15 @@ class OnDelivery implements Method
 	private $wp;
 	/** @var array */
 	private $options;
+    /** @var OrderServiceInterface */
+    private $orderService;
 
-	public function __construct(Wordpress $wp, Options $options)
-	{
-		$this->wp = $wp;
-		$this->options = $options->get('payment.'.self::ID);
-	}
+    public function __construct(Wordpress $wp, Options $options, OrderServiceInterface $orderService)
+    {
+        $this->wp = $wp;
+        $this->options = $options->get('payment.' . self::ID);
+        $this->orderService = $orderService;
+    }
 
 	/**
 	 * @return string ID of payment method.
@@ -108,6 +112,7 @@ class OnDelivery implements Method
 	public function process($order)
 	{
 		$order->setStatus(Order\Status::PROCESSING, __('Payment on delivery.', 'jigoshop'));
+        $this->orderService->save($order);
 
 		return '';
 	}